+++ docs/scripts/footer.js
@@ -1,5 +1,0 @@
-fetch("footer.html").then(function (response) {
-  response.text().then(function (text) {
-    document.getElementById("footer").innerHTML = text;
-  });
-});