+++ changelog.md
@@ -5,6 +5,7 @@
 - Add depth-first binary tree (dwindle)
 - Add insert into active window
 - Double buffer active windows
+- Updating docs (wip)
 
 ## 1.1